refactor handleLinks into a util
