+++ src/RouterSelector.web.js
@@ -2,8 +2,6 @@
 import React, { memo, useEffect, useState } from 'react'
 import { Platform } from 'react-native'
 import { pick } from 'lodash'
-import bip39 from 'bip39-light'
-import AsyncStorage from './lib/utils/asyncStorage'
 
 // components
 
@@ -22,65 +20,17 @@ import DeepLinking from './lib/utils/deepLinking'
 import InternetConnection from './components/common/connectionDialog/internetConnection'
 import isWebApp from './lib/utils/isWebApp'
 import logger from './lib/logger/pino-logger'
-import { APP_OPEN, fireEvent, initAnalytics, SIGNIN_FAILED, SIGNIN_SUCCESS } from './lib/analytics/analytics'
-import restart from './lib/utils/restart'
-
-// hooks
-import { savePathToStorage } from './lib/hooks/useSavePathToStorage'
-
-const { useSavePathToStorage } = savePathToStorage()
+import { APP_OPEN, fireEvent, initAnalytics } from './lib/analytics/analytics'
+import handleLinks from './lib/utils/handleLinks'
 
 const log = logger.child({ from: 'RouterSelector' })
 
 const DisconnectedSplash = () => <Splash animation={false} />
 
-/**
- * handle in-app links for unsigned users such as magiclink and paymentlinks
- * magiclink proceed to signin other links we keep and pop once user is logged in
- *
- * @returns {Promise<boolean>}
- */
-const handleLinks = async () => {
-  const params = DeepLinking.params
-
-  try {
-    const { magiclink } = params
-
-    let userNameAndPWD = Buffer.from(decodeURIComponent(magiclink), 'base64').toString()
-    let userNameAndPWDArray = userNameAndPWD.split('+')
-    log.debug('recoverByMagicLink', { magiclink, userNameAndPWDArray })
-    if (userNameAndPWDArray.length === 2) {
-      const userName = userNameAndPWDArray[0]
-      const userPwd = userNameAndPWDArray[1]
-      const UserStorage = await retryImport(() => import('./lib/gundb/UserStorageClass')).then(_ => _.UserStorage)
-
-      const mnemonic = await UserStorage.getMnemonic(userName, userPwd)
-
-      if (mnemonic && bip39.validateMnemonic(mnemonic)) {
-        const mnemonicsHelpers = retryImport(() => import('./lib/wallet/SoftwareWalletProvider'))
-        const { saveMnemonics } = await mnemonicsHelpers
-        await saveMnemonics(mnemonic)
-        await AsyncStorage.setItem('GD_isLoggedIn', true)
-        fireEvent(SIGNIN_SUCCESS)
-        restart('/')
-      }
-    }
-  } catch (e) {
-    if (params.magiclink) {
-      log.error('Magiclink signin failed', e.message, e)
-      fireEvent(SIGNIN_FAILED)
-    } else {
-      log.error('parsing in-app link failed', e.message, e, params)
-    }
-  }
-}
-
 let SignupRouter = React.lazy(async () => {
-  const { params } = DeepLinking
   const [module] = await Promise.all([
     retryImport(() => import(/* webpackChunkName: "signuprouter" */ './SignupRouter')),
-    handleLinks(),
-    useSavePathToStorage(params, log),
+    handleLinks(log),
     delay(animationDuration),
   ])
 
@@ -159,7 +109,7 @@ const RouterSelector = () => {
     setCheckedForBrowserSupport(true)
   }, [isLoggedIn])
 
-  // statring anumation once we're checked for browser support and awaited
+  // starting animation once we're checked for browser support and awaited
   // the user dismissed warning dialog (if browser wasn't supported)
   return (
     <React.Suspense fallback={<Splash animation={checkedForBrowserSupport} isLoggedIn={isLoggedIn} />}>
